pyproject.toml
[project]
name = "documentor"
version = "0.1.0"
description = "Library for parsing and semantic data recovery from documents of various formats"
authors = [{ name = "angrymuskrat", email = "mlhakov2011@gmail.com" }]
readme = "README.md"
license = { text = "BSD-3-Clause" }
requires-python = ">=3.10"

# Зависимости: заменяем ^ на PEP 440-диапазоны; убираем backport-пакеты для 3.10+
dependencies = [
    "pandas>=2.0.3,<3.0.0",
    "openpyxl>=3.0.10,<4.0.0",
    "scikit-learn>=1.3.2,<2.0.0",
    "numpy>=1.25.2,<2.0.0",
    "plotly>=5.17.0,<6.0.0",
    "natasha>=1.6.0,<2.0.0",
    "wikipedia2vec>=2.0.0,<3.0.0",
    "nltk>=3.8.1,<4.0.0",
    "torch>=2.4.0,<3.0.0",
    "torchvision>=0.19.0,<0.20.0",
    "matplotlib>=3.9.1,<4.0.0",
    # OCR dependencies
    "openai>=1.0.0,<2.0.0",
    "pillow>=10.0.0,<11.0.0",
    "pypdf2>=3.0.0,<4.0.0",
    "langchain-core>=0.1.0,<1.0.0",
    "overrides>=7.0.0,<8.0.0",
    "pdf2image (>=1.17.0,<2.0.0)",
]

[project.optional-dependencies]
jupyter = [
    "jupytext>=1.16.1,<2.0.0",
    "jupyter>=1.0.0,<2.0.0",
]

[tool.poetry]
requires-poetry = ">=2.0"

[tool.poetry.group.dev.dependencies]
jupytext = "^1.16.1"
jupyter = "^1.0.0"
pytest = "^8.0.1"
sphinx-rtd-theme = "^2.0.0"
sphinx = ">=5,<8"
sphinx-markdown-builder = "^0.6.7"

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"

